backend/requirements.txt
fastapi
uvicorn
pydantic>=2.0
pytest
httpx
sqlalchemy
databases
psycopg2-binary
redis
python-multipart
email-validator
pydantic_ai
openai
python-dotenv
googletrans
passlib[bcrypt]
deep-translator
pyjwt
bcrypt==4.0.1
passlib
python-jose[cryptography]
pillow==11.1.0

